+++ src/preprocessors/cpp.rs
@@ -1,11 +1,9 @@
 use crate::moduleloaders::{FileSystemModuleLoader, HttpModuleLoader};
 use crate::new_greco_rt_builder2;
+use gpp::{process_str, Context};
 use quickjs_runtime::eserror::EsError;
 use quickjs_runtime::esruntime::{EsRuntime, ScriptPreProcessor};
 use quickjs_runtime::esscript::EsScript;
-use quickjs_runtime::esvalue::EsValueConvertible;
-use quickjs_runtime::quickjs_utils::get_global_q;
-use quickjs_runtime::quickjs_utils::objects::get_property_q;
 use std::sync::Arc;
 
 lazy_static! {
@@ -64,53 +62,26 @@ impl ScriptPreProcessor for CppPreProcessor {
 
         log::debug!("CppPreProcessor > {}", script.get_path());
 
-        let rt: Arc<EsRuntime> = UTIL_RT.clone();
+        let src = script.get_code();
 
-        let init_res: Result<(), EsError> = rt.exe_rt_task_in_event_loop(|q_js_rt| {
-            // todo create specific context?
-            let q_ctx = q_js_rt.get_main_context();
-            let global = get_global_q(q_ctx);
-            let obj = get_property_q(q_ctx, &global, "CppPreProcessor")?;
-            if obj.is_null_or_undefined() {
-                // init func
-                q_ctx
-                    .eval(EsScript::new(
-                        "CppPreProcessor.not_es",
-                        "this.process = {};\n\
-                            this.CppPreProcessor = {process: function(src, vars){\n\
-                            let compiler_lib = require('https://raw.githubusercontent.com/ParksProjets/C-Preprocessor/master/lib/compiler.js');\n\
-                            let compiler = new compiler_lib.Compiler();\n\
-                            compiler.createConstant('DEBUG', 'true');\n\
-                            let res = {};\n\
-                            compiler.on('success', function(result) {\n\
-                                res.result = result;\n\
-                            });\n\
-                            compiler.on('error', function(e) {\n\
-                                res.err = e;\n\
-                            });\n\
-                            compiler.compile(src);\n\
-                            if (res.result) {\n\
-                                return(res.result);\n\
-                            } else {\n\
-                                throw Error(res.err);\n\
-                            }\n\
-                        }};",
-                    ))?;
-            }
-            Ok(())
+        let mut ctx = Context::new();
 
-        });
-
-        if init_res.is_err() {
-            return Err(EsError::new_string(format!(
-                "cpp preproc failed: {}",
-                init_res.err().unwrap()
-            )));
+        #[cfg(debug_assertions)]
+        {
+            ctx.macros.insert("DEBUG".to_string(), "true".to_string());
+        }
+        #[cfg(test)]
+        {
+            ctx.macros.insert("TEST".to_string(), "true".to_string());
+        }
+        #[cfg(not(any(debug_assertions, test)))]
+        {
+            ctx.macros.insert("RELEASE".to_string(), "true".to_string());
         }
 
-        let src = script.get_code().to_string().to_es_value_facade();
-        let new_code = rt.call_function_sync(vec!["CppPreProcessor"], "process", vec![src])?;
-        script.set_code(new_code.get_str().to_string());
+        let res = process_str(src, &mut ctx).map_err(|e| EsError::new_string(format!("{}", e)))?;
+
+        script.set_code(res);
         Ok(script)
     }
 }
@@ -126,9 +97,9 @@ mod tests {
         let res = rt.eval_sync(EsScript::new(
             "test.es",
             "((function(){\n\
-        #if 'DEBUG' == 'false'\n\
+        #ifdef HELLO\n\
             return 111;\n\
-        #elif 'DEBUG' == 'true'\n\
+        #elifdef DEBUG\n\
             return 123;\n\
         #else\n\
             return 222;\n\