+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 0.0.2 (work in progress)
 
-* ifdef preprocessor
+* ifdef preprocessor, uses [gpp](https://github.com/Kestrer/gpp)
 * macro preprocessor
 
 0.0.1